[ANDROID-135] fix : 목표 완료 퍼센테이지를 double 타입으로 전달해주고있었다

+++ core/data/src/main/java/see/day/network/dto/goal/CurrentGoalResponse.kt
@@ -15,7 +15,7 @@ data class CurrentGoalResponse(
     @Serializable(with = FlexibleDateTimeArraySerializer::class)
     val endDate: String,
     val completedDays: Int,
-    val achievementRate: Int,
+    val achievementRate: Double,
     val treeStage: TreeStage,
     val canCreateNew: Boolean
 )
